# feat: implement performance optimizations for Issue #58