--- a/crates/maos-core/Cargo.toml
+++ b/crates/maos-core/Cargo.toml
@@ -20,9 +20,6 @@ tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 parking_lot = { workspace = true }
 flate2 = { workspace = true }
-jsonschema = { workspace = true }
-schemars = { workspace = true }
-dirs = { workspace = true }
 
 [dev-dependencies]
 # Unit test dependencies (fast, no I/O)